Feat(procedural-map-gen): Got the refactored map generator working.
I finished refactoring the procedural map generator tool so that it now compiles and
will prompt the user for the size of the map to generator and print out the generated map.
The next step is to return some JSON and possibly enable this to work for a 3d map, as well.